Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main(){
-    float a,b;
-    scanf("%f",&a);
-    scanf("%f",&b);
+    double a,b;
+    scanf("%lf",&a);
+    scanf("%lf",&b);
     product = a*b;
     printf("Product: %.2f",product);
 }
